fix: blur on the top of the card and blur on the ticker
