+++ .changeset/lovely-papayas-scream.md
@@ -1,0 +1,5 @@
+---
+"live-mobile": minor
+---
+
+fix: blur on the top of the card and blur on the ticker
